requirements.txt
flask>=3.0.0
pillow>=11.0.0
werkzeug>=3.0.1
numpy>=1.26.0
google-generativeai
mongoengine
flask-login
pyopenssl
cryptography
authlib
requests
gunicorn
